Crate cranelift_object

Top-level lib.rs for cranelift_object.

This re-exports object so you don’t have to explicitly keep the versions in sync.

Re-exports

pub use object;

Structs

ObjectBuilder
A builder for ObjectModule.

ObjectModule
An ObjectModule implements Module and emits “.o” files using the object library.

ObjectProduct
This is the output of ObjectModule’s finish function. It contains the generated Object and other information produced during compilation.

Constants

VERSION
Version number of this crate.